Refactoring and optimization of RestoreTableCommand  * RestoreTableCommand moved to org.apache.spark.sql.delta.commands package  * cache() of filesToRemove DataFame removed (according to https://github.com/delta-io/delta/pull/863#issuecomment-1015672532)  * cache() of filesToAdd will be applied only if spark.sql.files.ignoreMissingFiles = false (default value)
