+++ core/src/main/scala/io/delta/tables/execution/DeltaTableOperations.scala
@@ -17,9 +17,8 @@
 package io.delta.tables.execution
 
 import scala.collection.Map
-
-import org.apache.spark.sql.delta.{DeltaLog, RestoreTableCommand}
-import org.apache.spark.sql.delta.commands.{DeltaGenerateCommand, VacuumCommand}
+import org.apache.spark.sql.delta.DeltaLog
+import org.apache.spark.sql.delta.commands.{DeltaGenerateCommand, RestoreTableCommand, VacuumCommand}
 import org.apache.spark.sql.delta.util.AnalysisHelper
 import io.delta.tables.DeltaTable
